Update clientes.component.ts
mensaje de confirmación para eliminar cliente

+++ src/app/cliente/clientes-table/clientes.component.ts
@@ -5,7 +5,7 @@ import { ClientesService } from '../../core/service/clientes.service';
 import { FormClienteComponent } from '../form-cliente/form-cliente.component';
 import { AutenticacionService } from 'src/app/usuarios/autenticacion.service';
 import { ActivatedRoute } from '@angular/router';
-
+import swal from 'sweetalert2';
 
 @Component({
   selector: 'app-clientes',
@@ -18,7 +18,6 @@ export class ClientesComponent implements OnInit {
   clientes: Cliente[] = [];
   paginador: any;
 
-
   constructor(
     private clientesService: ClientesService,
     private dialog: MatDialog,
@@ -28,7 +27,6 @@ export class ClientesComponent implements OnInit {
 
 
   ngOnInit(): void {
-    // this.allClientes();
     this.clientesPaginables();
    
   }
@@ -46,12 +44,10 @@ export class ClientesComponent implements OnInit {
         response => {
           this.clientes = response.content as Cliente[];
           this.paginador = response;
-          
         }
       )
     });
   }
-
 
   openDialog(dec: number, id: number){
     const dialogRef = this.dialog.open(FormClienteComponent);
@@ -62,19 +58,33 @@ export class ClientesComponent implements OnInit {
     } )
   }
 
-  allClientes(){
-    this.clientesService.getClientes().subscribe(r =>{
-      this.clientes = r
-      console.log(this.clientes);
-      
-    } );
-  }
-
   deleteCliente(id: number){
-    this.clientesService.deleteCliente(id).subscribe(response =>{
-      console.log(response);
-      this.clientesPaginables();
+    swal.fire({
+      title: '¿Estás seguro?',
+      text: `Estás eliminando al usuario con ID ${id}, no podrás revertir esta acción!`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, Eliminar!',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed)
+      {
+        this.clientesService.deleteCliente(id).subscribe(response =>{
+         
+          this.clientesPaginables();
+          swal.fire(
+                'Cliente Eliminado!',
+                `Eliminado con Exito.`,
+                'success'
+              )
+            }
+          )
+      }
     })
+    
+
   }
 
   showAction(){
